Code Review - mawilliams7
Added suggestions to help improve code quality. Consider implementing all non-function logic in a main method so that it is easier for the reader to understand. Also consider adding function docstrings so that a reader is able to understand what a particular function does in the context of this code.